Introduce test based on setup/teardown infrastructure
