+++ package.json
@@ -186,6 +186,8 @@
     "integration-test-run-all": "mocha-isolated --pass-through-aws-creds --skip-fs-cleanup-check --max-workers=20 \"test/integration/**/*.test.js\"",
     "integration-test-run-basic": "mocha test/integrationBasic.test.js",
     "integration-test-run-package": "mocha-isolated --skip-fs-cleanup-check test/integrationPackage/**/*.tests.js",
+    "integration-test-setup-infrastructure": "node ./scripts/test/setup-integration-infra.js",
+    "integration-test-teardown-infrastructure": "node ./scripts/test/teardown-integration-infra.js",
     "lint": "eslint .",
     "lint:updated": "pipe-git-updated --ext=js -- eslint",
     "pkg:build": "node ./scripts/pkg/build.js",
