+++ scripts/test/kafka.server.properties
